feat: 跟随 EasyOnebot
- Lgr 命名空间改名 Nc: `GroupPoke` / `FriendPoke` / `SendGroupForwardMsg` / `SendPrivateForwardMsg`
- 发送结果改用 `Easyonebot.SendMsgResult`, 去掉了对 `api.SendAnyMsgResp` 的依赖
- `MessageId` (int) 统一为 `MessageID` (int64), 同步 `biliParseTrackSet` / `registerTimerRecall` 与各调用点

+++ utilsMangaSearch.go
@@ -452,10 +452,10 @@ pica://%s`,
 	)
 }
 
-func registerTimerRecall(msgId int) {
+func registerTimerRecall(msgId int64) {
 	go func() {
 		<-time.After(RECALL_DURATION)
-		onebot.Call().Std.DeleteMsg(msgId)
+		onebot.Call().Std.DeleteMsg(int(msgId))
 	}()
 }
